+++ src/datasource/DroneDatasource.ts
@@ -95,14 +95,14 @@ export class DroneDatasource {
     page?: any,
     search?: string,
     sortField?: string,
-    sortDirection?: string
+    sortDirection?: string,
   ): Promise<DroneBrandListEntity> {
     const params = {
       take: take,
       page: page,
       search: search,
       sortField: sortField,
-      sortDirection: sortDirection
+      sortDirection: sortDirection,
     }
     return httpClient
       .get(BASE_URL + '/drone-brand/droner-brand-all', { params })